Fixed tests, named `Coso` and fixed websocket url resolution
